Upload cp.py
This is the updated file cp.py and is now the main file for the Crypto-Prophet Application.  Streamlit.py as of this commit is an older version that does not include the features that the cp.py file has.
Features included in this file are Top 100 by Marketcap, Coin Analysis, Deep Analysis, Cycle Analysis, Google Trends, and Tweet Count, although Tweet count is not working. This is the file we will use for the presentation 11/18/21 in class.